@solnic 2015-01-09
Nice! But...if it mutates then it should be renamed to `:group!`. It would also be nice to find some more lower-level transformations that are used inside (kinda like with wrap that uses nest).