@splattael 2015-01-09
After `dup`ing `:group` transproc it's still faster than the old impl.

```
$ ruby -Ilib group.rb  
Calculating -------------------------------------
             group 1     3.749k i/100ms
         group old 1     3.696k i/100ms
-------------------------------------------------
             group 1     43.305k (± 0.7%) i/s -    217.442k
         group old 1     42.134k (± 3.5%) i/s -    210.672k

Comparison:
             group 1:    43305.5 i/s
         group old 1:    42133.8 i/s - 1.03x slower

Calculating -------------------------------------
            group 10   642.000  i/100ms
        group old 10   551.000  i/100ms
-------------------------------------------------
            group 10      6.600k (± 0.9%) i/s -     33.384k
        group old 10      5.660k (± 0.9%) i/s -     28.652k

Comparison:
            group 10:     6600.0 i/s
        group old 10:     5660.1 i/s - 1.17x slower

Calculating -------------------------------------
           group 100    89.000  i/100ms
       group old 100    72.000  i/100ms
-------------------------------------------------
           group 100    898.357  (± 0.7%) i/s -      4.539k
       group old 100    724.435  (± 1.8%) i/s -      3.672k

Comparison:
           group 100:      898.4 i/s
       group old 100:      724.4 i/s - 1.24x slower

Calculating -------------------------------------
          group 1000     9.000  i/100ms
      group old 1000     7.000  i/100ms
-------------------------------------------------
          group 1000     91.433  (± 1.1%) i/s -    459.000 
      group old 1000     72.612  (± 2.8%) i/s -    364.000 

Comparison:
          group 1000:       91.4 i/s
      group old 1000:       72.6 i/s - 1.26x slower
```

I will look at low-level transformation soonish.