@splattael 2015-03-26
Mh, I can't even remember this PR oO

Thanks for merging! :green_heart: